MGS-2027 [#ccc] Integrate VSI 6.1.1 gpu driver release
New features:
1. Wayland GBM backend support.
2. Multi-GPU combined mode and individual mode support.
3. Multi-GPU affinity control interface.
4. Linux DMA-BUF buffer sharing to support Chrome OS.

Date: Jul 19, 2016

+++ driver/compiler/libCLC/llvm/tools/clang/lib/Basic/Diagnostic.cpp
@@ -103,7 +103,7 @@ static const StaticDiagInfoRec *GetDiagInfo(unsigned DiagID) {
 #endif
 
   // Search the diagnostic table with a binary search.
-  StaticDiagInfoRec Find = { DiagID, 0, 0, 0, 0, 0, 0 };
+  StaticDiagInfoRec Find = { (unsigned short)DiagID, 0, 0, 0, 0, 0, 0 };
 
   const StaticDiagInfoRec *Found =
     std::lower_bound(StaticDiagInfo, StaticDiagInfo + NumDiagEntries, Find);
